PP-5256 Organised classes and minor changes
